Simulator + tests: build fixes after layout/virtual-WAS rebase
- MainWindowViewModel: disambiguate Math.Clamp(short,int,int) vs
  Math.Clamp(int,int,int) by explicit (int) cast on the input value.
- IntegrationTests.csproj: re-apply GenerateProgramFile=false so the
  existing Program.Main coexists with the test SDK's auto-Main.
  Lost when this branch rebased onto develop's older csproj.

+++ Simulators/AgValoniaGPS.VehicleSimulator/ViewModels/MainWindowViewModel.cs
@@ -244,7 +244,7 @@ public class MainWindowViewModel : INotifyPropertyChanged
         get => _virtualWasOffset;
         set
         {
-            _virtualWasOffset = (short)Math.Clamp(value, -2000, 2000);
+            _virtualWasOffset = (short)Math.Clamp((int)value, -2000, 2000);
             OnPropertyChanged();
             if (_hub != null) _hub.Steer.VirtualWasOffset = _virtualWasOffset;
         }
